2ND WORKING VERSION, NEEDS SOME FINISHES
