+++ src/ca/sheridancollege/project/Game.java
@@ -6,6 +6,7 @@
 package ca.sheridancollege.project;
 
 import java.util.ArrayList;
+import java.util.Scanner;
 
 /**
  * The class that models your game. You should create a more specific child of this class and instantiate the methods
@@ -51,7 +52,7 @@ public abstract class Game {
     /**
      * Play the game. This might be one method or many method calls depending on your game.
      */
-    public abstract void play();
+    public abstract void play(Scanner sc);
 
     /**
      * When the game is over, use this method to declare and display a winning player.
